Get feedback and stuff working for drive

+++ EmbeddedTeam13/firmware/src/encoders.h
@@ -10,10 +10,10 @@ int getRightEncoderCount(void);
 int getLeftEncoderCountISR(void);
 int getRightEncoderCountISR(void);
 
-int getLeftEncoderSpeed(void);
-int getRightEncoderSpeed(void);
-int getLeftEncoderSpeedISR(void);
-int getRightEncoderSpeedISR(void);
+float getLeftEncoderSpeed(void);
+float getRightEncoderSpeed(void);
+float getLeftEncoderSpeedISR(void);
+float getRightEncoderSpeedISR(void);
 
 void lEncoderIsr();
 void rEncoderIsr();
